Module Prime_enummap

Enumerated map (prime.unstable).

This is similar to Map from the standard library, but provides a locate method which return the position of a binding within a map.

module type OrderedType = sig ... end

module type S = sig ... end

module type S_monadic = sig ... end

module type S_with_monadic = sig ... end

module Make : functor (Key : OrderedType) -> sig ... end

module Make_monadic : functor (Key : OrderedType) -> functor (Monad : Prime_sigs.Monad) -> S_with_monadic with type key = Key.t and type 'a monad = 'a Monad.t