Module Prime_enumset

Enumerating set (prime.unstable).

This is similar to the Set structure from the standard library, but provides a locate method to tell the position of an element in the map.

module type OrderedType = sig ... end

module type S = sig ... end

module type S_monadic = sig ... end

module type S_with_monadic = sig ... end

module Make : functor (Elt : OrderedType) -> sig ... end

module Make_monadic : functor (Elt : OrderedType) -> functor (Monad : Prime_sigs.Monad) -> S_with_monadic with type elt = Elt.t and type 'a monad = 'a Monad.t