Module Prime_wip__Prime_free_magma

This module implements a free magma over an arbitrary type. This is in effect a binary tree with values stored on the leafs and no balancing. It can be used as an alternative to a list when efficient repeated concatenation is needed.

type 'a t =

| Gen of 'a

| Cat of 'a t * 'a t

val gen : 'a -> 'a t

val cat : 'a t -> 'a t -> 'a t

val rev : 'a t -> 'a t

val fold : ('a -> 'b -> 'b) -> 'a t -> 'b -> 'b

val fold_rev : ('a -> 'b -> 'b) -> 'a t -> 'b -> 'b

val to_seq : 'a t -> 'a Stdlib.Seq.t

val to_list : 'a t -> 'a list